+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Armies with multiple commanders are more likely to spawn on multiple worlds if possible
 - Upgrade Techs
   - Advanced Fabrication Aircraft
+  - Advanced Fabrication Bot
 
 ### Fixed